requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.7.14
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
crispy-bootstrap4==2025.6
Django==5.1.3
django-allauth==65.3.0
django-crispy-forms==2.4
django-jsonfield-backport==1.0.5
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-notifs==4.0.0
django-qr-code==4.1.0
django-recurrence==1.11.1
django-simple-history==3.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.10
jsonfield==3.2.0
packaging==25.0
pillow==11.0.0
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.4
segno==1.6.1
setuptools==80.9.0
six==1.16.0
sqlparse==0.5.2
swapper==1.4.0
typing_extensions==4.12.2
urllib3==2.5.0
